# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta API Json Rest actua como proxy SOAP a los servicios web de ARCA.",
        "title": "API proxy a los webservices de ARCA",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Sebastian Hogas",
            "email": "sehogas@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/gestabref/ConsultarFechaUltAct": {
            "get": {
                "description": "Retorna la fecha de última actualización de la tabla consultada.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consulta de Tablas de Referencia"
                ],
                "summary": "Obtener la Fecha de última actualización de la tabla",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de referencia",
                        "name": "IdReferencia",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wgestabref.FechaUltAct"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gestabref/Dummy": {
            "get": {
                "description": "Visualizar el estado del servicio web, del servicio de autenticación y de la base de datos de ARCA",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consulta de Tablas de Referencia"
                ],
                "summary": "Muestra el estado del servicio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wgestabref.WsDummyResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gestabref/ListaArancel": {
            "get": {
                "description": "Retorna tabla del tipo código / descripción / opción / vigencia.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consulta de Tablas de Referencia"
                ],
                "summary": "Lista Arancel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de referencia",
                        "name": "IdReferencia",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wgestabref.Opciones"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gestabref/ListaDatoComplementario": {
            "get": {
                "description": "Lista Datos Complementarios",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consulta de Tablas de Referencia"
                ],
                "summary": "Lista Datos Complementarios",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de referencia",
                        "name": "IdReferencia",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wgestabref.DatosComplementarios"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gestabref/ListaDescripcion": {
            "get": {
                "description": "Emite tabla del tipo código / descripción.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consulta de Tablas de Referencia"
                ],
                "summary": "Lista Descripción",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de referencia",
                        "name": "IdReferencia",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wgestabref.Descripciones"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gestabref/ListaDescripcionDecodificacion": {
            "get": {
                "description": "Lista Descripción Decodificación",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consulta de Tablas de Referencia"
                ],
                "summary": "Lista Descripción Decodificación",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de referencia",
                        "name": "IdReferencia",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wgestabref.DescripcionesCodificaciones"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gestabref/ListaEmpresas": {
            "get": {
                "description": "Emite tabla del tipo cuit / razón social.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consulta de Tablas de Referencia"
                ],
                "summary": "Lista de Empresas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de referencia",
                        "name": "IdReferencia",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wgestabref.Empresas"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gestabref/ListaLugaresOperativos": {
            "get": {
                "description": "Emite tabla del tipo código / descripción / vigencia / aduana / lugar operativo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consulta de Tablas de Referencia"
                ],
                "summary": "Lista de Lugares Operativos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de referencia",
                        "name": "IdReferencia",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wgestabref.LugaresOperativos"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gestabref/ListaPaisesAduanas": {
            "get": {
                "description": "Emite tabla del tipo código / descripción / vigencia /país o aduana.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consulta de Tablas de Referencia"
                ],
                "summary": "Lista de Paises y Aduanas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID de referencia",
                        "name": "IdReferencia",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wgestabref.PaisesAduanas"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/gestabref/ListaTablasReferencia": {
            "get": {
                "description": "Emite tabla del tipo: Tabla de Referencia / Descripción Tabla Referencia / WebMethod (que se debe utilizar para obtener los datos de dicha tabla).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Consulta de Tablas de Referencia"
                ],
                "summary": "Lista de Tablas de Referencia",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key de acceso",
                        "name": "x-api-key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wgestabref.TablasReferencia"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "wgestabref.DatosComplementarios": {
            "type": "object",
            "properties": {
                "CodError": {
                    "type": "integer"
                },
                "DatosComplementarios": {
                    "$ref": "#/definitions/wgestabref.ArrayOfDatoComplementario"
                },
                "IdReferencia": {
                    "type": "string"
                },
                "InfoAdicional": {
                    "type": "string"
                }
            }
        },
...
        "wgestabref.Descripciones": {
            "type": "object",
            "properties": {
                "CodError": {
                    "type": "integer"
                },
                "Descripciones": {
                    "$ref": "#/definitions/wgestabref.ArrayOfDescripcion"
                },
                "IdReferencia": {
                    "type": "string"
                },
                "InfoAdicional": {
                    "type": "string"
                }
            }
        },
...
        "wgestabref.FechaUltAct": {
            "type": "object",
            "properties": {
                "CodError": {
                    "type": "integer"
                },
                "Fecha": {
                    "type": "string"
                },
                "IdReferencia": {
                    "type": "string"
                },
                "InfoAdicional": {
                    "type": "string"
                }
            }
        },
...
        "wgestabref.Opciones": {
            "type": "object",
            "properties": {
                "CodError": {
                    "type": "integer"
                },
                "IdReferencia": {
                    "type": "string"
                },
                "InfoAdicional": {
                    "type": "string"
                },
                "Opciones": {
                    "$ref": "#/definitions/wgestabref.ArrayOfOpcion"
                }
            }
        },
...
        "wgestabref.WsDummyResponse": {
            "type": "object",
            "properties": {
                "appserver": {
                    "type": "string"
                },
                "authserver": {
                    "type": "string"
                },
                "dbserver": {
                    "type": "string"
                }
            }
        },
...
    }
}